requirements.txt
aiogram==3.13.0
yt-dlp
python-dotenv
